+++ test/soda4.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved. */
+/* Copyright (c) 2018, 2020, Oracle and/or its affiliates. All rights reserved. */
 
 /******************************************************************************
  *
@@ -36,7 +36,7 @@ const testsUtil = require('./testsUtil.js');
 describe('168. soda4.js', () => {
 
   before(async function() {
-    const runnable = await testsUtil.checkPrerequisites();
+    const runnable = await testsUtil.isSodaRunnable();
     if (!runnable) {
       this.skip();
       return;